refactor: transform test/connect to ts
